chore: small refactor of the file name
