+++ packages/angular/projects/sdk/package.json
@@ -21,15 +21,15 @@
   ],
   "peerDependencies": {
     "@angular/common": "^19.0.0 || ^20.0.0 || ^21.0.0",
-    "@angular/core": "^19.0.0 || ^20.0.0 || ^21.0.0",
+    "@angular/core": "^19.0.0 || ^20.0.0 || ^21.0.0"
+  },
+  "dependencies": {
+    "tslib": "^2.8.1",
     "@storyblok/live-preview": "workspace:*",
     "storyblok-js-client": "workspace:*",
     "@storyblok/richtext": "workspace:*",
     "@storyblok/js": "workspace:*",
     "@storyblok/preview-bridge": "^2.1.6"
-  },
-  "dependencies": {
-    "tslib": "^2.8.1"
   },
   "sideEffects": false
 }
